improve no migrations found messages
