+++ base/commands/migration/status_stages.go
@@ -4,13 +4,10 @@ package migration
 
 import (
 	"context"
-	"encoding/json"
-	"fmt"
 
 	"github.com/hazelcast/hazelcast-commandline-client/clc/ux/stage"
 	"github.com/hazelcast/hazelcast-commandline-client/internal/plug"
 	"github.com/hazelcast/hazelcast-go-client"
-	"github.com/hazelcast/hazelcast-go-client/serialization"
 )
 
 type StatusStages struct {
@@ -29,11 +26,13 @@ func (st *StatusStages) Build(ctx context.Context, ec plug.ExecContext) []stage.
 			FailureMsg:  "Could not connect to the migration cluster",
 			Func:        st.connectStage(ec),
 		},
+		{
+			ProgressMsg: "Finding migration in progress",
+			SuccessMsg:  "Found migration in progress",
+			FailureMsg:  "Could not find a migration in progress",
+			Func:        st.findMigrationInProgress(ec),
+		},
 	}
-}
-
-type MigrationInProgress struct {
-	MigrationID string `json:"id"`
 }
 
 func (st *StatusStages) connectStage(ec plug.ExecContext) func(context.Context, stage.Statuser[any]) (any, error) {
@@ -43,24 +42,16 @@ func (st *StatusStages) connectStage(ec plug.ExecContext) func(context.Context, 
 		if err != nil {
 			return nil, err
 		}
+		return nil, nil
+	}
+}
+
+func (st *StatusStages) findMigrationInProgress(ec plug.ExecContext) func(context.Context, stage.Statuser[any]) (any, error) {
+	return func(ctx context.Context, status stage.Statuser[any]) (any, error) {
 		m, err := findMigrationInProgress(ctx, st.ci)
 		if err != nil {
 			return nil, err
 		}
 		return m.MigrationID, err
 	}
-}
-
-func findMigrationInProgress(ctx context.Context, ci *hazelcast.ClientInternal) (MigrationInProgress, error) {
-	var mip MigrationInProgress
-	q := fmt.Sprintf("SELECT this FROM %s WHERE JSON_VALUE(this, '$.status') IN('STARTED', 'IN_PROGRESS', 'CANCELING')", StatusMapName)
-	r, err := querySingleRow(ctx, ci, q)
-	if err != nil {
-		return mip, fmt.Errorf("finding migration in progress: %w", err)
-	}
-	m := r.(serialization.JSON)
-	if err = json.Unmarshal(m, &mip); err != nil {
-		return mip, fmt.Errorf("parsing migration in progress: %w", err)
-	}
-	return mip, nil
 }
